Return miss and issue delete to NvmCache to on refcount overflow
Summary:
Initially my plan was to propogate back the error but I realize it presents some usability problems. Hence I decided to return miss and issue delete.

Today, with only DRAM, we would get a refcount exception when we call CacheAllocator::find() and the refcount overflows. However, with NvmCache enabled, we would get a miss instead.

If we propogate the error back to the user, then we can either throw a refcount exception when user calls "get()" or simply expose it via a flag like "wasRefcountOverflow()". Both are quite different from what we currently guarantee which is to throw when calling "CacheAllocator::find()".

So since this is expected to happen very rarely, I've opted to simply return a miss to the first reader that caused the overflow and all subsequent readers, and also issue delete for the key.

Reviewed By: sathyaphoenix

Differential Revision: D18922596

fbshipit-source-id: a9e1d90e697c0966e7c757721afda153a011f767