RFC: Unified stats structure
Summary:
With the interface to two counter types unified, this diff unifies the
stats structure. This enables us to
1) keep the relevant stats for different components together. Ex nvm stats are
all next to each other.
2) choose the appropriate method of thread-safety and atomicity based on the
stat and not what the popular group requirement is. Now we can have some nvm
stats using atomic and some using thread local.

That said, this is not trivial. So if you have any reservations about this
methodology, please comment so. There is also an issue of stack overflow in
here because we have too many std::array of MaxClass x MaxPool. With
ThreadLocal, each member of array is 96 bytes.  So I have resorted to moving
those arrays to be heap allocated, resulting in additional dereferences.

Reviewed By: therealgymmy

Differential Revision: D19060407

fbshipit-source-id: 88e4d4153afc2e129a172d6c5fbba4324e0cf5d5