Introduce TLCounter wrapper around FastStats
Summary:
Provides same interface as AtomicCounter so that we can merge the counters
into one struct and use them uniformly. Today, for example, we have some nvm
counters as TLStats instead of atomic because it is convienient to group them
under one category instead of bifurcating. Having the same interface makes the
intent expressed when we declare the stat and all places using both types of
counters use them with apis.

Reviewed By: therealgymmy

Differential Revision: D18994937

fbshipit-source-id: e2134494ef271d16e22410ec769a276845ea83b6