Update Navy's hole stats correctly
Summary:
Previously, when we remove the index entry during reclaim, we could remove the wrong item (which is why since it's treated as a removal). This diff makes sure we only remove an item when the address is the same.

In addition, expose the value replaced from Index::insert(), so we can update the hole stats accordingly as well.

Reviewed By: sathyaphoenix

Differential Revision: D18991217

fbshipit-source-id: fcafa9cc01dfa1553fce10b2856dd9bba0b866ab