Add PieceWiseReplayGenerator
Summary:
Add the basic logic in PieceWiseReplayGenerator. detailed piecewise caching and range request is not implemented in this diff yet.

The class maintains activeReqM_ for ongoing request. getReq() gets a request from the trace
and add it into activeReqM_; notifyResult() removes the request from activeReqM_ when it's
cache hit or set success.

Reviewed By: sathyaphoenix

Differential Revision: D19456346

fbshipit-source-id: cdfbee1a9fa2607adf80ad4ee3cd444685d573b3