Move Rust bindings for ReadOnlySharedCacheView into separate crate
Summary: To address the binary size issues outlined in T58951274, our decl service client needs to avoid linking to cachelib. {D18923363} provides a thin C++ library to do so. This diff moves the Rust bindings for ReadOnlySharedCacheView out of the main Rust cachelib bindings crate into a new crate which links to the thin ro-cache-view library.

Reviewed By: ljw1004

Differential Revision: D19323216

fbshipit-source-id: 7407bc1db4c07093874e9a68f778558c3b062a09