Changed the message in CacheAllocator::startNewWorker to DBG1 instead of INFO on success.
Summary:
This message gets printed to stdout in lots of binaries.
IMO success should generally be silent so I think that a lower log level
makes sense. It looks like the log was added to fix a task that has since been closed. Maybe we can just delete it completely? Sending to the person who added the log as the one best qualified to judge. :)

Reviewed By: therealgymmy

Differential Revision: D19742200

fbshipit-source-id: 1b9a50f0cb37f7b3c9222de5c2cc41dc530a7533