Deprecate RegionManager lock in favor of Region's internal locks
Summary:
Moved lock inside Region and renamed methods to reflect the change.

The functional change is that in Allocator::allocate() code path, the "pinning" or "setting a class id" is now done as a separate atomic operation than "opening and allocate" from a region. But that shouldn't affect the correctness as RegionAllocator's lock still synchronizes them correctly. We still maintain the invariant "one region is associated with exactly one region allocator when it is in use for writes".

Reviewed By: sathyaphoenix

Differential Revision: D19422969

fbshipit-source-id: db027ef1151473969d621ce64739ace9ba9ebd20