Refactor stressor by passing generator instance
Summary:
cachebench is a synthetic concurrent stressor and validator for CacheLib. The benchmark has a workload generator component that generates api requests that can be played on cache. The current workload generator has evolved to a messy state and can be simplified.

Today, workload generators are a template parameters to the CacheStressor. this is no longer needed and the workload generators can be instantiated and used through their virtual interface.

Reviewed By: sathyaphoenix

Differential Revision: D20043511

fbshipit-source-id: 2dd09b84396aba42dc01374a8118f90749efb7da