Navy in-mem buffer support for read/writes
Summary:
Navy in-mem buffer support for read/writes

Added support to do buffered writes and reads to/from navy. Writes to navy would write to a buffer and return and when the buffer is full, it is flushed to the device. Similarly, reads from navy are read from in-memory buffer if present.

This feature is controlled by "dipper_navy_num_in_mem_buffers" option.

Reviewed By: sathyaphoenix

Differential Revision: D18374192

fbshipit-source-id: 5c9ceb530f872f3420f4a44f691aef1b0007f668