Bound test duration by a user-specified maximize runtime
Summary:
Basically we should let a cachebench test finish when it has finished all its test operations or it has spent longer than "maximum allowed time". This should prevent any of our tests from timing out anymore on cogwheel. We can have a 6 hours timeout that is the maximum amount of time we can run on Cogwheel. If we get past that time, we can shut down successfully (passing the test).

The default timeout value is set to 0 to make outside users free to control their tests.

Reviewed By: therealgymmy

Differential Revision: D20127393

fbshipit-source-id: 641b7d4cd73e1eade7ca316ce508fb805475cd3b