Support chained items in sampling and/or Scuba
Summary:
Support chained items in sampling and/or Scuba.
I'm not married to the naming here.  Also, if you think code should
move between files, let me know.  I hope the locking is OK.
I was surprised that iterating the chain didn't seem to include the
parent.  Let me know if it's supposed to.  Also, there was mention in
the task about validating the parent key, but I didn't see what to do.

The unit test exercises the code, but I'm open to improvements that
will make it more strict.

Reviewed By: sathyaphoenix

Differential Revision: D20109741

fbshipit-source-id: 50add28ac74ac7dcadf9d9fcc4b069029c1baffa