shard the piecewise replay generation logic
Summary:
We have a single mutex protecting all the request generation. This causes
contention and blocks us from increasing the throughput.

This diff splits the locks into two. A sharded lock for active request handling and a separate lock for getline concurrency.  This also converts some of the stats to atomic counters so that the critical sections for the lock are smaller.

Reviewed By: therealgymmy, fxzhang

Differential Revision: D20452076

fbshipit-source-id: ead096fc4556eabada3e3ad76c85f90f013fc8da