Use spin lock instead of std::mutex
Summary:
Trying to improve the throughput of replay from navy side quickly makes the
replay generator become the bottleneck.  Worse, increasing the threads in
replay generator to send more load to navy causes replay generator to be slow.

The use of std::mutex was becoming a hot contention point in notifyResult.
Simply switching to a spin lock saves the trip to kernel space given that this
is highly contended with a small critical section.

also, convert some of the lookups to be single lookup per call instead of a
count + find.

Reviewed By: fxzhang

Differential Revision: D20449441

fbshipit-source-id: 3f08b4dabd5fe270281a9a64099a29da333f1fd8