Reinforce R/W Access Mode (1/n)
Summary:
Backgroud NvmCache provide API find(key, mode) that returns a ItemHandle. The ItemHandle should be a writable only if If mode=write. And the user then can call getWritableMemory() from it to get a writable memory. If mode=read, then the ItemHandle should be readonly, and getWritableMemory() will raise exceptions. The client should only call getMemory() to get a 'const T*' so any inconsistent can be found at compile time.

This diff do one things:
1. Add the new getWritableMemory() API

The following diffs are about to:
1) Inside cachelib, convert all improper getMemory() to getWritableMemory()
2) Implement AccessMode-getWritableMemory consistency check
3) Evict from nvm when find(key, writeMode) is called
4) Convert all improper callsites within fbcode to getWritableMemory()
5) Change getMemeory() return type to `const void*` to prevent misusage in the future.

Reviewed By: sathyaphoenix

Differential Revision: D20376719

fbshipit-source-id: 541c6aea8ef5939ab99fda9a200a0ff9f1b8cbd8