Benchmark BTree and F14FastMap Speed & Memory Footprint
Summary:
Create benchmark test to compare cachelib BTree with F14FastMap, in terms of speed and memory usage. The results are pasted.

For speed, F14FastMap is **16x** faster, which shows a lot of CPU benefit to use a hashmap instead of the BTree.

However, BTree wins on memory footprint. As shown in the results, F14FastMap use 26% more memory.

Conclusion: We should keep looking for good hashmap solution that maybe a little bit slower but with better memory effieciency. Now in my mind tsl::sparse_map is a good candidate, and I'll add it into the benchmark in the next diff.

Note: std::unordered_map doesn't provide an easy way to get the memory stats, but it's notorious memory usage and bad performance just save me that effort.

Reviewed By: sathyaphoenix

Differential Revision: D20667017

fbshipit-source-id: 1ab295f21b83130d4cee948d0c394a08554b8459