Update nvme_helper tool to avoid using shell when running nvme
Summary: This diff updates the NandWrites code to invoke the `nvme` tool directly, instead of through `nandWriteBytes`, so that we avoid using the shell. We need to avoid using the shell because we want to use a setuid root tool to read the NVMe NAND write totals when running in a non-root context, but bash specifically drops privileges if it's invoked by a setuid root binary (as a security measure).

Reviewed By: sathyaphoenix

Differential Revision: D19916704

fbshipit-source-id: 5b1c17e4b3a584047873cc8048bcd6e724f460c4