fix online generator stuck on registerThread()
Summary:
The root cause is the reqs_ size checking in OnlineGenerator<Distribution>::registerThread(). The only exit condition is `reqs_.size() == config_.numThreads`, aka 36, but when in prepolulation, the actual number of threads may be config_.prepopulateThreads (aka 50), so the extra 14 threads has no way to match the exit contidion.

The fix is:
1) Totally remove config_.prepopulateThreads
2) For online generator, define key_ and req_ as folly::ThreadLocal, so we can kill the keys_ and reqs_ map, as well as the mutex
3) After 2), we could finially kill registerThread()

Reviewed By: sathyaphoenix

Differential Revision: D20909026

fbshipit-source-id: 9a0640d8624d29520b9534f806b6c596c71334fa