relax hashtable bit size to not be power of two
Summary:
hashtable bit size was restricted to be a power of two to support modulo
arithmetic. Switching it to be not restrictive since modulo is a small part of
the overall computation. This enables us to arrive at better bounds for large
filters which are hundreds of mb in size.

Reviewed By: therealgymmy

Differential Revision: D21014170

fbshipit-source-id: ce302323e9c5aaf4269db517993516dcd1e1a91f