stop relying on bloom filter init bits
Summary:
Going back, looking at the history(D9928000,  D15189570), we used the init bit
in the bloom filter to indicate when bighash was recovered but bloom filters
were not persisted.  Now that we have support to persist bloom filters, the
init bits serve no purpose to get/set. They are always initialized and only
cleared briefly when we are rebuilding the filter for a bucket on
insert/remove

I think the previous code changes to add support for persisting bloom filters
missed cleaning this part up. However, would be good to confirm if this
understanding is correct. If this looks good, following diff will simplify the
implementation of bloom filter

Reviewed By: therealgymmy

Differential Revision: D20973528

fbshipit-source-id: 2b81978cd8fb2353cf71e5df60a30c1b45bdeda1