Navy region should check for in-memory buffer before declaring it ready for reclaim
Summary:
Navy region should check for in-memory buffer before declaring it ready for reclaim

While running cachebench tests with in memory buffers enabled, a condition has occurred that a region with memory buffer is evicted before the in memory buffer is flushed. The flush happens later by which time the region has been re-allocated and there are writers are actively writing to the region.
Sequence of events -
1. Thread-1 starts flushing a region. Puts region in pending flush state but does not complete flush because  active reader/writers is not 0
2. Thread-2 picks the same region for eviction and by this time active reader/writer is 0 and successfully evicts
3. Thread-3 picks the evicted region and starts using it. Active reader/writer is non-zero
4. Thread-1 will successfully flush the region when reader/writer count becomes zero buffer_ is released
5. Thread-3 tries to use it again but buffer_ is null now

Reviewed By: therealgymmy

Differential Revision: D21386483

fbshipit-source-id: 6218ad628318ff56727e762d1320effc58045421