Kill prepopulation feature
Summary:
Prepopulate started off with the goal of making it possible to get reflective throughput
numbers once cache is warmed up. Perhaps a reasonable trade-off is to run this
longer or fix the stats collection to not begin until a certain timeframe.

The way prepopulation is special cased in the stressor has been ad-hoc and it
is getting harder to make it work across many stressor use cases.  Few
scenarios:
* harder to do thread local logic with workload generators like in D20929911
* prepopulate considers only dram evictions and the current logic already by
  passes that with other stuff like continuous hits.

Perhaps a good trade-off is to simply consider running the workload for longer
to get the noise from cache warm up stage disregarded or provide a way to get
throughput per sampling interval.

for now, killing the current implementation and revisting this holistically
makes sense.

Reviewed By: fxzhang

Differential Revision: D21541803

fbshipit-source-id: b650ef394fd4b4f70195dc35c47c1557c80bbf9e