Cachelib changes: Add a second parameter to weight function
Summary:
For the weighted LruTailAge strategy, we plan to use a weight function described as follow:

weight = (( classId + 1 ) * coefficient ) / numAllocationClasses

Currently, the weight function passed to LruTailAge strategy accepts 1 parameter only: ClassId.

Now, we need to add a second parameter to the weight function, so that we can provide the number of allocation classes as described above.

Reviewed By: sathyaphoenix

Differential Revision: D21583394

fbshipit-source-id: ae3d0e6026e59e34f255b1afead08d5963abb3b4