Implement NewMoveCb for Cachelib clients
Summary:
Introduce `NewMoveCb` as `function<Item&, Item&, std::optional<Item*>>`. The extra optional pointer is the parent item for the item being moved, if the item is a chained item. This would help callsites in move callback that needs to perform operation on parents for chained items.

`NewMoveCb` and `MoveCb` would coexist in the code base while we migrate from `MoveCb` to `NewMoveCb`. We will deprecate `MoveCb` (and renaming the new one) once migration finishes in the near future.

Reviewed By: therealgymmy

Differential Revision: D21760174

fbshipit-source-id: da12e7cca3fe2f647761d8ec4e42e05ff8d64c66