Navy read during reclaim throws due to pread error
Summary:
Navy read during reclaim throws due to pread error.

The failure is a result of two diffs D21999505 and D22004275 that landed separately.

One diff has changed the read size but did not check the return value to the correct size. The other diff made to throw error when size mismatch occurs.

Reviewed By: sathyaphoenix

Differential Revision: D22093353

fbshipit-source-id: 5d1cd64262692eab9da6ad14ebd60b79d23713af