Properly align down device size for RAID0Device
Summary:
This will ensure we will never end up with a device with insufficient space at the end for a Stripe. We will gate this behind an option that is by default true, but disabled for BigCache.

For BigCache, the plan is to release this via configerator flag which will require another diff in BigCache code.

Reviewed By: sathyaphoenix

Differential Revision: D22244158

fbshipit-source-id: 58c583979650a9b31e2345432f27ca6b74693665