Always delete cold roll file if exists on ShmManager creation
Summary:
This got triggered in TAO because we have two caches. When RW cache is cold, we cold roll regular cache as well. When that happens, we skip calling "SharedMemAttach", and as a result we do not clean up "ColdRoll" file. This means on the next restart, we'll again have a cold roll.

This diff changes the behavior to always delete the ColdRoll for ShmManager when we detect the ColdRoll file.

Reviewed By: vrishal

Differential Revision: D22652348

fbshipit-source-id: ee29ffe08dc7350b35b788c09493b7957a57077a