make default reaper to be slab-walk mode
Summary:
This changes the reaper configuration for those that use reaper currently to
switch to slabwalk mode. All other configuration such as interval and
throttler stay the same.

With this change,  the reaper thread will sleep for the said interval as
before, however, it would only sleep after scanning the cache completely.
While scanning, it will throttle itself based on the config passed.

A recommended way to get around would be to set the throttler config passed to
the reaper with the same sleep interval as the original reaper.

Reviewed By: therealgymmy

Differential Revision: D22519127

fbshipit-source-id: e42d2cf4f66840d30ef29e494b7001a386a91b4c