remove support for iterator based reaper
Summary:
to simplify the reaper configurations, we are consolidating on using the slab
walk reaper instead of the iterator based reaper. The slab walk reaper is more
efficient since it scans memory for items that are potentially expired, before
grabbing locks to confirm and expire.

this change cleans up the iterator based reaper. a following diff will
simplify the cache allocator config api and clean up the cachelib user code
that sets this up.

Reviewed By: therealgymmy

Differential Revision: D23037020

fbshipit-source-id: e27d9555c02e3213b191e6743065e9ea5d6b32e6