enforce alignment with the segment's size instead of config size
Summary:
SlabAllocator internally needs a Slab aligned start address to do pointer
arithmetic. To accomplish this, we mmap a region that is sufficiently large
and re-map an aligned address into this mapping that we control. this ensures
the re-map behavior is safe. Except when cache size changes, the mapped region
might be smaller than the cache size. This can lead to vm maps of other
mappings to  be hijacked, causing random corruptions. ASAN detects this, but
gets confused pointing to what exactlt the problem is. In some cases,
syscalls segfault. So, while we expect an exception to be thrown from size
mismatch, we hit random corruption.

To fix this, we now do the temporary mapping closer to the actual mapping,
with the appropriate size of the segment. If cache is sized down, the re-map
is safe and we would throw an exception as expected from the SlabAllocator
restoration checks.

Reviewed By: therealgymmy

Differential Revision: D23150549

fbshipit-source-id: e1e8ba9c2f7a196025414000785988578c0684f1