don't wait until evictions to reap
Summary:
Reaper originally had this option to avoid causing lock contention and only
doing the reaping until we were out of memory. The current approach also does
not wait for all pools to be full, but rather starts reaping when any of them
is full, which is incomplete from the original intention. I think it is
simpler to just avoid this configuration and have reaper be freeing expired
items as they expire instead.

Reviewed By: therealgymmy

Differential Revision: D23194227

fbshipit-source-id: 0a5f447fe602c747ee5b86a2245a4895d50d1499