Deprecate concept of "free" regions in Navy BlockCache
Summary:
As title. This avoid a source of potential bugs. We only have "free" regions when Navy is freshly started up. Once Navy has used all the regions and running in steady state, we never have "free" regions.

This change gets rid of this concept, and tracks all regions on startup. This puts Navy into the steady state logic right away. The very first insertion will trigger an eviction, and so on.

Reviewed By: sathyaphoenix

Differential Revision: D22861100

fbshipit-source-id: c6af1a1ba57094fbc9b44e8037aa35eedd2ac48e