Update MockEvictonPolicy to default to FIFO
Summary:
As title. Default MockEvictionPolicy to using the FIFO policy. Only keep expecations on the policy if they're meaningful in the unit tests.

Also fix a bug with tracking a region on recovery. Previously it wouldn't work for a FIFO policy as "touch" has no effect. Now we first track unused regions and then track used regions.

Reviewed By: sathyaphoenix

Differential Revision: D23202128

fbshipit-source-id: 1fc4ba88c256a5b19ee17d58e30b40bdb3398502