Fix incorrect casting for OpType in workload generator
Summary:
I introduced this bug originally with the switch to "util::narrow_cast". This made the cast from uint8_t to OpType result in value which is always "0" which is set.

This should fix it and return correct operation types.

This issue only affects the default workload generator and does not affect online, trace replay, and piecewise generators.

----

Separately, I fixed an unrelated bug that WorkloadGenerator and Distribution objects keep a const StressorConfig& member which refers to a copy of the config object created on stack in `makeStressor`. It hasn't caused any issue because we don't directly use anything from the config object, but it's a bug nonetheless.

Reviewed By: sathyaphoenix

Differential Revision: D23644087

fbshipit-source-id: 5d0576f367ba76cd6be793bc0f419dbbd2c5f59a