Remove permanent item support in NVM
Summary:
Part of the plan in deprecating permanent items from CacheLib. Since all permanent items never get evicted from RAM cache, we can safely remove permanent item support in NVM cache, which was built when RAM cache was not persistent between rolls and is no longer useful.

The assumption is that all the item that ever inserted into NVM will be evictable. If NVM ever receives a permanent item from RAM, we will log that and treat it as non-permanent.

On the spirit of backward compatibility, I still keep the proto field isPinned but it will always be set to false.

Once this diff lands, the next step will be to remove the stat once we are confident.

Reviewed By: therealgymmy

Differential Revision: D23243836

fbshipit-source-id: dec5c0db5abc34402b9db9069a5f85753b9c3a23