Verify alignment changes correctly when in-mem buffer is enabled
Summary:
For >16TB devices, we error out on startup for BigCache because we check in config validation that we have sufficient address bits (up to 32 bit) to address the entire flash drive.

However, with in-mem buffer feature turned on for BigCache, we actually automatically adjust alignment size to make sure we can address the entire flash drive. So the config check is actually NOT necessary.

This diff updates it to check for the correct alignment.

In addition, we only check for BigHash bucket alignment if BigHash is enabled.

Reviewed By: shodoco

Differential Revision: D23715360

fbshipit-source-id: 44a062340900d08ee0291161ef878cd3092a43c4