Serialize the precedure an allocation class claims a new slab
Summary: MemoryPool::allocate(int size) is not serialized, we may run into the problem that multiple allocations to the same allocation class were not met and they all asked for free slabs and then attach the free slabs to the allocation class. This will result in the allocation class having a lot of free slabs.

Reviewed By: sathyaphoenix

Differential Revision: D23763008

fbshipit-source-id: 79870edf5999152e901046826937fa534b05b000