Compute weights based on an exponential weight function OR a weight function that looks into a map with classId to weight data.
Summary:
Per Memcache Architecture Evolution document (https://fb.quip.com/P2q8AcM2Uyb2), for the weighted slab rebalancer, we want to have greater control over the length of any Lru.

In order to do this, we have enabled a way to set a unique weight per allocation class, and this is achieved by having "a new weight function" that looks into a map with classId to weight data, and returns corresponding weight. This will give us complete control per classId granularity.

The following is an example of the map. The key for each entry in the map represents a class Id, and the corresponding value represents the weight it is mapped to.

```
{{ 1 : 0.22 },
 { 2 : 0.45 },
 { 3 : 0.98 }
 <more>
 { 120: 1.0}}
```

* This "new weight function" receives AllocInfo, and then it looks up into the map using the AllocInfo.classId as key, and then returns corresponding weight value. If the classId is not found in the map, it returns a weight of "1.0".

* The weights to set in the map are yet to be determined as more experimentation and tests are performed.

* The original weight function is not longer a lineal function, but an exponential function which receives a classId, and it no longer receives number of allocation classes as second parameter. Again, this may change in the future as more tests are performed in canary/shadow or some prod tiers.

* Memcache/cachetable controls passing the exponential weight function, OR the weight function that looks into a map, to the slab rebalancer. These two weight functions are mutually exclusive.

Note: This diff covers changes in cachelib, which will be the base for the diffs in memcache/cachetable(D23290378) and configerator(D22880688)

We should weight the lower allocation class LRUs heavier than the higher allocation classes, this would increase the number of items being served from DRAM, and therefore lowering latency. This would also decrease flash burn-rate as a collateral.

Reviewed By: sathyaphoenix

Differential Revision: D23192878

fbshipit-source-id: 86f392b19de9bdefe1bd8a65a20a0a0a07b9eb83