Make sure cache size change due to alignment will be warm roll forward and backward
Summary:
The way this stack works is as follows:

1. Roll out D22772014 (this diff). It's a no-op.
2. Roll out D22543493 (this ensures cache size is aligned to region size). This is warm-roll forward and backward.
3. Roll out D22772400. This cleans up the code. This will no longer be compatible with older version (cache size mismatch).

This diff readies the code to make the alignment change for cache size to be warm roll forward and backward for all Navy use cases. We added logic that if we detect Navy's version is "12" or higher, but previous version was 11, then we will handle change in "cache size", as long as the "cache size  / region size" is the same between the two configs. And also the same idea vice-versa.

Reviewed By: sathyaphoenix

Differential Revision: D22772014

fbshipit-source-id: 2a34fb7cbc9a7dad9d40a5aa0a22ecc5fa314bb6