Implement SFIFO policy
Summary:
Refer to `1/n` for overall design.

This diff introduces a SegmentedFIFO policy that tracks a region in a segment that matches its class id.

For example, if we have 3 segments, we expect the regions to come from 3 classes {0, 1, 2}, where 0 is the lowest priority, and will be evicted first.

This policy supports rebalancing between the different segments. But the rebalancing is one direction only (from highest priority to lowest priority).   This means the lowest-pri segment can grow to far larger than its ratio suggests. This is okay, as we only need higher-pri segments for items that are deemed important

Reviewed By: sathyaphoenix

Differential Revision: D23861337

fbshipit-source-id: 5fe0e95e078a5e851030da3fa1d98afb4d1cb6d4