Make online generator not be limited by key space
Summary:
Previously, when we use online generator, we would prepopulate the ~4B
accesses from the popularity distribution and replay that over and over. This
did not generate same popularity when scaling the #keys and cacheSizeMb to
compare scenarios. Ideally, sccaling these two equally should produce same hit
ratio. however, due to the 4billion prepopulated popularity, the distributions
were distinct that larger configurations never even evicted.

this diff fixes that by doing the following
-- use 32bit->64bit rng
-- instead of prepopulating the distributions, call into the popularity
distribution directly. There is some run time cost, but it is okay given the
in-accuracy and the fact that we are already doing online request generation.

Reviewed By: therealgymmy

Differential Revision: D24441459

fbshipit-source-id: b3da605fd30b79b40418af453b9f8b7ac2fe9a4e