Use size field in Navy index
Summary:
As title. This diff starts using the size field in Navy's index. It will ensure that when running in stacked mode, we read only as much as we need. This should bridge latency gap between stacked mode vs. size class for TAO, and also improve latency for BigCache.

It has the following changes:
1. On insertion, we will save a 16 bit size hint into the index.
2. It's warm-roll compatible, if Navy detects an entry has no size field, it will simply defer back to "readBuffer_"

Note that because we store a 16 bit size hint, the size which we decode from the hint may be slightly bigger than the real size (aligned up to 512 bytes which is our minimal alloc size alignment when using in-memory buffer).

This seems a bit convoluted for now, but once we deprecate other write modes. The 512 bytes alignment will be obvious.

Reviewed By: sathyaphoenix

Differential Revision: D24182757

fbshipit-source-id: 6d484f8dd13d308dc9ba173de683f060bac6c68c