Use updateTime as ramEvictionAgeSecs
Summary:
I want a metric that gives me a length of time window that I can slide on cache accesses traces to get flash admissions (or ram evictions).

The updateTime here is defined as the last time the item is updated in the MMcontainer (LRU queue). So I think it is the appropriate to use.

Reviewed By: therealgymmy

Differential Revision: D24811266

fbshipit-source-id: cc43824fd46d1de4a871e6e4e7cb294bf58e110c