Always use O_DIRECT and detect incompatibility
Summary:
We expect users to always pass O_DIRECT except when they pass a file that
could be on tmpfs which bails on O_DIRECT due to historic reasons. Instead of
having this behavior specified by user who has no clue about this, we detect
if the path is on tmpfs and try to use O_DIRECT appropriately.

Reviewed By: therealgymmy

Differential Revision: D24662516

fbshipit-source-id: 2f54f0b2e056cd32c6e088b4d159e14b85d012ae