Remove use of template for workload generators to simplify configuration
Summary:
Workload generators that needed a popularity distribution had to be templated
to accomodate different implementation types. This made the configuration of
cachebench hard to explain to users. Now, with this cleanup, users don't need
to specify the distribution type using a string identifier. Instead they can
configure the popBuckets and popWeights to indicate if we are using discrete
popularity or if we are using normal distribution.

Will fix the documentation  at https://www.internalfb.com/intern/wiki/Cache_Library_User_Guides/Configuring_cachebench_parameters/#popularity-and-size-dist once approved.

Reviewed By: therealgymmy

Differential Revision: D25896322

fbshipit-source-id: 531046e1f0e2cb566781d1f9c8f1bf18283e2589