Fix Navy DynamicRandomAP probabilityFactor cumulation
Summary:
We didn't apply the capping correctly. So if we are going 1.25x every 60 seconds, it is very easy for this double to inf. Once it goes to double::inf, it won't come back with any multiplier.

To validate this: With target rate at 34MB/s, I ran a script that calls the update with a write rate of 20MB/s for 50000 minutes (under writing, so probFactor multiply 1.25x every minute). Then calls update with write rate 100MB/s for 50000 minutes (overwriting, so probFactor should decrease by 0.75x every minute). The factor stayed at double::inf.
But if I first do 50000 minutes of 100MB/s then 50000 minutes of 20MB/s, it stays at 99.88e-324 (basically 0).

This change does two things:
1. Apply a range on probabilityFactor. It will always stay in [0.001, 10]. So in the worst case, it takes 9 minutes for rate limiting to take into effect (10 * 0.75^9 = 0.75), and it takes 31 minutes for the host to resume traffic from a heavily rate limited pattern (0.01 * 1.25^31 = 1.00).
2. (WS-only) Change the base probability calculation to base on WS regions size. Since WS has only one item size, it makes sense for everyone to receive a baseProbability of 1.

Reviewed By: sathyaphoenix

Differential Revision: D25938455

fbshipit-source-id: 2754596bdb88211d76654a086daf9f9d8dce9098