Re-allocate handle when parent has changed while moving
Summary:
In one of the recent improvements for moving logic, we re-designed the logic
to not allocate under the moving sync and to hold the moving sync for a
smaller critical section. This was done to reduce contention on user level
locks that are passed as moving syncs. One of the downsides now is that when
we allocate a new item for moving a chained item, the parent could be
replaced.

To handle this, we check under the moving sync for the pre-condition and abort
if the condition is not met. The retry logic will try allocating again and
after sufficient attempts if it fails, will resort to evicting. This should be
rare.

Reviewed By: therealgymmy

Differential Revision: D26005327

fbshipit-source-id: eb28ab250bcd087645f201884513e202724bb3bc