Add a hard-cap write rate to DynamicRandom
Summary: DynamicRandom was designed to only take into account device write budget and set a target rate around that. When a host receives little traffic for extensive period of time and suddenly starts receiving overwhelming load, the ap wouldn't reject because based on budget calculation the incoming rate is acceptable but unrealistic from a device standpoint.

Reviewed By: sathyaphoenix

Differential Revision: D26002975

fbshipit-source-id: 392179c9e66faf1d91b4a99bccb37f8b301e528b