Break down trackAccess and getAccesses in AccessTracker
Summary:
As part of unifying training, we are going to make sure we do track the access at cache level and get the accesses at flash admission. This is also required to migrate BigCache to use NvmAdmissionPolicy.

This change is the first step, to break down the two functionality into different public functions for AccessTracker. For now, WS and TAO still uses the old function, which calls into these two.

Reviewed By: sathyaphoenix

Differential Revision: D26188348

fbshipit-source-id: 2f22f014b1ea52face920f08c8173349eb669095