Suppress data races in cachelib/allocator/memory/SlabAllocator.cpp
Summary:
This suppresses some data races reported by TSAN. See the associated
task(s) below for context, including sample stack traces caused by these races
and reproduction instructions.

This diff is automatically generated. Therefore, the way it makes suppressions
may not be as beautiful as if written by hand. *However, we don't have the
resources to manually adjust these diffs, nor do we have the capacity to
actually fix the bugs*; we just want to get the existing bugs
out of the way so we can enable TSAN across the fleet. If you are a reviewer
please do one of the following:

1. Accept the diff as is, and you may follow up with more changes (or fix the
   bugs) later.
2. Fix the data races in a different diff and land it within a reasonable amount
   of time (e.g. a week), and comment about it here.
3. Comment to suggest us a different code location(s) to suppress these data
   races.

Reviewed By: igorsugak

Differential Revision: D26094374

fbshipit-source-id: ebd6019da6690dd35340b86054e3e7b50ee870eb