Suppress a series of data races in cache allocator reaper.
Summary:
Also commented in the associated Task:

This is a series of data races caused by having incompletely initialized data structures that are readable from the reaper thread as soon as their memory is allocated.

For example, as soon a CacheItem is allocated, the reaper has access to it and will try to query whether it has expired, even before the constructor of that CacheItem is called. That isn't very well-defined behavior, and in particular, the constructor that writes the expiryTime_ would race with the reaper thread. This apparently happens with the slab header as well.

Reviewed By: sathyaphoenix

Differential Revision: D26519476

fbshipit-source-id: 2d1f0df74b1da096362243726041e31c26e73a52