fix flay schedluer test
Summary:
When writing the test, we didn't consider  the raciness involved in producers
enqueueing and tracking queue length with consumers picking up the work and
waiting on the first job. We could have at most one job per queue being
executed and assumed instead that one job would always be currently executed.
This resulted in sandcastle marking the tests as flaky. Fix the test to account for these.

Reviewed By: therealgymmy

Differential Revision: D26639095

fbshipit-source-id: 7ba005fcb9662f4656feea26450655d990fbe8b2