Introducing CountDownLatch
Summary:
CountDownLatch is a synchronization mechanism to allows a thread to wait for other threads to finish before proceeding.
https://docs.oracle.com/javase/7/docs/api/java/util/concurrent/CountDownLatch.html

This will be available in C++20, but probably not anytime soon to fbcode. https://en.cppreference.com/w/cpp/thread/latch.

I built the interface to be similar to std::latch.

Reviewed By: therealgymmy

Differential Revision: D26123370

fbshipit-source-id: 2933e8ae668dfb9ec305bcfd8745c36ec01563ca