Introducing Ticker interface
Summary:
This diff introduces the Ticker interface for CacheLib components to get "ticks".

A tick is a unit in time. Currently, without formally defining it we are using seconds as a tick.
This diff replaces the getCurrentTick function in AccessTracker with the Ticker interface, which allows us overriding the time management in the future.

Reviewed By: therealgymmy

Differential Revision: D26193078

fbshipit-source-id: 51dd796e9f25d8e615c8b9eba279b405ae00be42